add SMI_AD3_Eventos 'Finalizar la app web de gestión de eventos'

+++ SMI_AD3_Eventos/src/main/java/eventos/modelo/entitys/Evento.java
@@ -2,7 +2,6 @@ package eventos.modelo.entitys;
 
 import java.io.Serializable;
 import jakarta.persistence.*;
-import java.math.BigDecimal;
 import java.util.Date;
 
 
@@ -43,7 +42,7 @@ public class Evento implements Serializable {
 
 	private String nombre;
 
-	private BigDecimal precio;
+	private double precio;
 
 	//uni-directional many-to-one association to Tipo
 	@ManyToOne
@@ -133,11 +132,11 @@ public class Evento implements Serializable {
 		this.nombre = nombre;
 	}
 
-	public BigDecimal getPrecio() {
+	public double getPrecio() {
 		return this.precio;
 	}
 
-	public void setPrecio(BigDecimal precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
